refactor(model)!: Split vulnerability classes to a separate package
For a better overview, split out CVSS classes to separate files in a new
"vulnerabilities" sub-package of the ORT model. A plural name for the
package was chosen to align with the "licenses" sub-package.
