+++ model/src/main/kotlin/config/VulnerabilityResolution.kt
@@ -19,7 +19,7 @@
 
 package org.ossreviewtoolkit.model.config
 
-import org.ossreviewtoolkit.model.Vulnerability
+import org.ossreviewtoolkit.model.vulnerabilities.Vulnerability
 
 /**
  * Defines the resolution of an [Vulnerability]. This can be used to silence false positives, or vulnerabilities that
