+++ model/src/main/kotlin/config/VulnerabilityResolutionReason.kt
@@ -19,7 +19,7 @@
 
 package org.ossreviewtoolkit.model.config
 
-import org.ossreviewtoolkit.model.Vulnerability
+import org.ossreviewtoolkit.model.vulnerabilities.Vulnerability
 
 /**
  * Possible reasons for resolving an [Vulnerability] using a [VulnerabilityResolution].
