+++ model/src/test/kotlin/config/VulnerabilityResolutionTest.kt
@@ -22,7 +22,7 @@ package org.ossreviewtoolkit.model.config
 import io.kotest.core.spec.style.WordSpec
 import io.kotest.matchers.shouldBe
 
-import org.ossreviewtoolkit.model.Vulnerability
+import org.ossreviewtoolkit.model.vulnerabilities.Vulnerability
 
 class VulnerabilityResolutionTest : WordSpec({
     "matches" should {
