+++ model/src/test/kotlin/vulnerabilities/VulnerabilityReferenceTest.kt
@@ -17,52 +17,12 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.model
+package org.ossreviewtoolkit.model.vulnerabilities
 
 import io.kotest.core.spec.style.StringSpec
-import io.kotest.matchers.nulls.beNull
-import io.kotest.matchers.should
 import io.kotest.matchers.shouldBe
 
-import org.ossreviewtoolkit.model.VulnerabilityReference.Cvss2Rating
-import org.ossreviewtoolkit.model.VulnerabilityReference.Cvss3Rating
-
-class VulnerabilityTest : StringSpec({
-    "The CVSS 2 rating should be correct for a given score" {
-        Cvss2Rating.fromScore(-0.1f) should beNull()
-
-        Cvss2Rating.fromScore(0.0f) shouldBe Cvss2Rating.LOW
-        Cvss2Rating.fromScore(3.9f) shouldBe Cvss2Rating.LOW
-
-        Cvss2Rating.fromScore(4.0f) shouldBe Cvss2Rating.MEDIUM
-        Cvss2Rating.fromScore(6.9f) shouldBe Cvss2Rating.MEDIUM
-
-        Cvss2Rating.fromScore(7.0f) shouldBe Cvss2Rating.HIGH
-        Cvss2Rating.fromScore(10.0f) shouldBe Cvss2Rating.HIGH
-
-        Cvss2Rating.fromScore(10.1f) should beNull()
-    }
-
-    "The CVSS 3 rating should be correct for a given score" {
-        Cvss3Rating.fromScore(-0.1f) should beNull()
-
-        Cvss3Rating.fromScore(0.0f) shouldBe Cvss3Rating.NONE
-
-        Cvss3Rating.fromScore(0.1f) shouldBe Cvss3Rating.LOW
-        Cvss3Rating.fromScore(3.9f) shouldBe Cvss3Rating.LOW
-
-        Cvss3Rating.fromScore(4.0f) shouldBe Cvss3Rating.MEDIUM
-        Cvss3Rating.fromScore(6.9f) shouldBe Cvss3Rating.MEDIUM
-
-        Cvss3Rating.fromScore(7.0f) shouldBe Cvss3Rating.HIGH
-        Cvss3Rating.fromScore(8.9f) shouldBe Cvss3Rating.HIGH
-
-        Cvss3Rating.fromScore(9.0f) shouldBe Cvss3Rating.CRITICAL
-        Cvss3Rating.fromScore(10.0f) shouldBe Cvss3Rating.CRITICAL
-
-        Cvss3Rating.fromScore(10.1f) should beNull()
-    }
-
+class VulnerabilityReferenceTest : StringSpec({
     "The severity string should be correct for a given CVSS 2 score" {
         VulnerabilityReference.getSeverityString("CVSS2", "1.0") shouldBe "LOW"
         VulnerabilityReference.getSeverityString("CVSSV2", "5.0") shouldBe "MEDIUM"
